+++ src/lib/signups.ts
@@ -37,7 +37,7 @@ export interface Signup {
     request?: string;
     status: string;
     paymentTotal: number;
-    token: string;
+    paymentMethodId: string;
     stripeCustomerId: string;
     created: {
         _seconds: number;